Various improvements
Simulation analysis:
- fix scripts for neat systems
- compress archives after conversion
- remove uncompressed npy files
- save first snapshot with avi
- add super remote target

Experimental analysis:
- sleep between analyses (for xvfb)
- rebuild benchmark when scripts change
- scale colors in NP images, make background black
- show hexagonal peaks in RDF plots
- scale g(r) plots to nanometers
- fix normalization of g(r) plots
- load archives if available
- only save archives and plots that are missing

Other:
- start report
- ignore unanalyzed SEM files and all PDF files

+++ exp/common.py
@@ -93,11 +93,14 @@ class SEMAnalysis():
     def filterimage(self, gaussian1=1, gaussian2=1, balance=[(5,75),(3,25)]):
         """ Try to balance brightness and smooth a bit. """
 
-        self.filtered = ndimage.gaussian_filter(self.img, gaussian1)
-        for b in balance:
-            self.filtered = balanceimage(self.filtered, b[0], b[1])
-        self.filtered = ndimage.gaussian_filter(self.filtered, gaussian2)
-        self.filtered = stretchimage(self.filtered)
+        if os.path.exists(self.ffiltered+".npy.bz2"):
+            self.filtered = numpy.load(bz2.BZ2File(self.ffiltered+".npy.bz2"))
+        else:
+            self.filtered = ndimage.gaussian_filter(self.img, gaussian1)
+            for b in balance:
+                self.filtered = balanceimage(self.filtered, b[0], b[1])
+            self.filtered = ndimage.gaussian_filter(self.filtered, gaussian2)
+            self.filtered = stretchimage(self.filtered)
 
     def thresholdimage(self, factor=1.1, erode=True):
         """ Thresholding with optional erosion. """
@@ -105,9 +108,12 @@ class SEMAnalysis():
         if not hasattr(self, "filtered"):
             self.filterimage()
 
-        self.threshold = self.filtered > factor*mahotas.otsu(self.filtered)
-        if erode:
-            self.threshold = pymorph.erode(self.threshold)
+        if os.path.exists(self.fthreshold+".npy.bz2"):
+            self.threshold = numpy.load(bz2.BZ2File(self.fthreshold+".npy.bz2"))
+        else:
+            self.threshold = self.filtered > factor*mahotas.otsu(self.filtered)
+            if erode:
+                self.threshold = pymorph.erode(self.threshold)
 
     def labelimage(self, threshold=None):
         """ Label regions in image corresponding to regional maxima. """
@@ -141,7 +147,11 @@ class SEMAnalysis():
         else:
             self.nps = pymorph.cwatershed(self.dist, self.labels, Bc=numpy.ones((3,3), dtype=bool))
 
-        self.coms = numpy.array(ndimage.center_of_mass(self.filtered, self.nps, range(1,self.nlabels+1)))
+        if os.path.exists(self.fcoms+".npy.bz2"):
+            self.coms = numpy.load(bz2.BZ2File(self.fcoms+".npy.bz2"))
+        else:
+            self.coms = numpy.array(ndimage.center_of_mass(self.filtered, self.nps, range(1,self.nlabels+1)))
+
         self.regcom = numpy.zeros(self.img.shape, dtype="uint8")
         for icom,com in enumerate(self.coms):
             self.regcom[round(com[0]),round(com[1])] = 1
@@ -150,9 +160,26 @@ class SEMAnalysis():
         
         if not hasattr(self, "coms"):
             self.watershedimage()
-        self.rdf, self.bins = radialdistribution(self.coms, self.img)
 
-    def plot(self):
+        # Load RDFs if available.
+        if os.path.exists(self.frdf+".npy.bz2"):
+            self.bins, self.rdf = numpy.load(bz2.BZ2File(self.frdf+".npy.bz2"))
+        else:
+            self.rdf, self.bins = radialdistribution(self.coms, self.img)
+
+        # Find the location of the first (highest) peak in the RDF.
+        wlen = 3
+        self.window = numpy.array([1,3,1], dtype=float)
+        self.window /= sum(self.window)
+        self.rdfsmooth = numpy.convolve(self.window, self.rdf, mode='valid')
+        self.rdfsmoothmax = numpy.argmax(self.rdfsmooth)
+        self.binmax = self.rdfsmoothmax + wlen/2
+        self.fitX = self.bins[self.binmax-1:self.binmax+2]
+        self.fitY = self.rdf[self.binmax-1:self.binmax+2]
+        self.fit = scipy.poly1d(scipy.polyfit(self.fitX,self.fitY,deg=2))
+        self.rdfmax = -self.fit[1]/(2*self.fit[2])
+
+    def plot(self, rdfpixelsize=1.0, rdfxunits="pixels"):
         """ Helper routine for plotting in this task. """
 
         pylab.figure(1)
@@ -162,42 +189,57 @@ class SEMAnalysis():
         pylab.imshow(pymorph.overlay(self.threshold, self.seeds))
 
         pylab.figure(3)
-        pylab.imshow(self.nps % 55)
+        pylab.imshow((self.nps % 20 + 5)*(self.nps>0), cmap=pylab.cm.spectral)
 
         pylab.figure(4)
         pylab.imshow(pymorph.overlay(self.img, self.regcom))
 
         pylab.figure(5)
-        pylab.plot(self.bins, self.rdf)
-        pylab.ylim([0.0, max(3.0,max(self.rdf))])
-        pylab.xlim([0, 50])
-        pylab.xlabel("distance [pixels]")
+        pylab.plot(self.bins*rdfpixelsize, self.rdf)
+        ymax = max(3.0,max(self.rdf))
+        pylab.ylim([0.0, ymax])
+        pylab.xlim([0, 50*rdfpixelsize])
+        pylab.xlabel("distance [%s]" %rdfxunits)
+        for peak,label in [(1,"$1$"), (numpy.sqrt(3),"$\sqrt{3}$"), (2,"$2$"), (numpy.sqrt(7),"$\sqrt{7}$")]:
+            x = peak*self.rdfmax*rdfpixelsize
+            pylab.axvline(x=x, ymin=0, ymax=ymax, linestyle='--', color='gray')
+            pylab.text(x, ymax*1.01, label, fontsize=15, horizontalalignment="center")
         pylab.grid()
         a = pylab.axes([0.60, 0.60, 0.25, 0.25], axisbg='y')
-        pylab.setp(a, xlim=(0,.2), xticks=[50,100,200], yticks=[1.0])
-        pylab.xlim([0,256])
-        pylab.plot(self.bins, self.rdf)
+        pylab.setp(a, xlim=(0,.2), xticks=map(int,[50*rdfpixelsize,100*rdfpixelsize,200*rdfpixelsize]), yticks=[1.0])
+        pylab.xlim([0,256*rdfpixelsize])
+        pylab.plot(self.bins*rdfpixelsize, self.rdf)
         pylab.grid()
 
     def savearchives(self):
         """ Helper routine for saving NumPy archives in this task. """
 
-        numpy.save(self.ffiltered+".npy", self.filtered)
-        numpy.save(self.fthreshold+".npy", self.threshold)
-        numpy.save(self.fnps+".npy", self.nps)
-        numpy.save(self.fcoms+".npy", self.coms)
-        numpy.save(self.frdf+".npy", [self.bins, self.rdf])
+        if not os.path.exists(self.ffiltered+".npy.bz2"):
+            numpy.save(self.ffiltered+".npy", self.filtered)
+        if not os.path.exists(self.fthreshold+".npy.bz2"):
+            numpy.save(self.fthreshold+".npy", self.threshold)
+        if not os.path.exists(self.fnps+".npy.bz2"):
+            numpy.save(self.fnps+".npy", self.nps)
+        if not os.path.exists(self.fcoms+".npy.bz2"):
+            numpy.save(self.fcoms+".npy", self.coms)
+        if not os.path.exists(self.frdf+".npy.bz2"):
+            numpy.save(self.frdf+".npy", [self.bins, self.rdf])
 
     def savefigures(self):
         """ Helper routine for saving figures from pylab in this task. """
 
-        pylab.figure(1)
-        pylab.savefig(self.ffiltered+".png")
-        pylab.figure(2)
-        pylab.savefig(self.fthreshold+".png")
-        pylab.figure(3)
-        pylab.savefig(self.fnps+".png")
-        pylab.figure(4)
-        pylab.savefig(self.fcoms+".png")
-        pylab.figure(5)
-        pylab.savefig(self.frdf+".png")
+        if not os.path.exists(self.ffiltered+".png"):
+            pylab.figure(1)
+            pylab.savefig(self.ffiltered+".png")
+        if not os.path.exists(self.fthreshold+".png"):
+            pylab.figure(2)
+            pylab.savefig(self.fthreshold+".png")
+        if not os.path.exists(self.fnps+".png"):
+            pylab.figure(3)
+            pylab.savefig(self.fnps+".png")
+        if not os.path.exists(self.fcoms+".png"):
+            pylab.figure(4)
+            pylab.savefig(self.fcoms+".png")
+        if not os.path.exists(self.frdf+".png"):
+            pylab.figure(5)
+            pylab.savefig(self.frdf+".png")
